app/controllers/__init__.py
from webargs.flaskparser import parser
from flask.ext.restful import abort


@parser.error_handler
def handle_request_parsing_error(err):
    """webargs error handler that uses Flask-RESTful's abort function to return
    a JSON error response to the client.
    """
    abort(422, errors=err.messages)
