requirements.txt
mysql-connector-python-rf==2.1.3
Flask-RESTful==0.2.12
Flask-SQLAlchemy=2.1
webargs=0.15.0
flask-log=0.1.0